[Refactor] #1371 アイテム強化に関する処理のファイルを全体的にフォルダ分割した
