+++ Hengband/Hengband/Hengband.vcxproj
@@ -273,10 +273,10 @@
     <ClCompile Include="..\..\src\load\item\item-loader-factory.cpp" />
     <ClCompile Include="..\..\src\load\monster\monster-loader-factory.cpp" />
     <ClCompile Include="..\..\src\load\player-class-specific-data-loader.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\abstract-weapon-enchanter.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\object-smith.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\smith-info.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\smith-tables.cpp" />
+    <ClCompile Include="..\..\src\object-enchant\weapon\abstract-weapon-enchanter.cpp" />
+    <ClCompile Include="..\..\src\smith\object-smith.cpp" />
+    <ClCompile Include="..\..\src\smith\smith-info.cpp" />
+    <ClCompile Include="..\..\src\smith\smith-tables.cpp" />
     <ClCompile Include="..\..\src\object-use\item-use-checker.cpp" />
     <ClCompile Include="..\..\src\object-use\use-execution.cpp" />
     <ClCompile Include="..\..\src\object-use\zaprod-execution.cpp" />
@@ -293,15 +293,15 @@
     <ClCompile Include="..\..\src\main-win\main-win-term.cpp" />
     <ClCompile Include="..\..\src\main-win\wav-reader.cpp" />
     <ClCompile Include="..\..\src\monster\monster-damage.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\apply-magic-amulet.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\abstract-protector-enchanter.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\apply-magic-boots.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\apply-magic-cloak.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\apply-magic-crown.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\apply-magic-gloves.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\apply-magic-helm.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\apply-magic-ring.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\apply-magic-shield.cpp" />
+    <ClCompile Include="..\..\src\object-enchant\others\apply-magic-amulet.cpp" />
+    <ClCompile Include="..\..\src\object-enchant\protector\abstract-protector-enchanter.cpp" />
+    <ClCompile Include="..\..\src\object-enchant\protector\apply-magic-boots.cpp" />
+    <ClCompile Include="..\..\src\object-enchant\protector\apply-magic-cloak.cpp" />
+    <ClCompile Include="..\..\src\object-enchant\protector\apply-magic-crown.cpp" />
+    <ClCompile Include="..\..\src\object-enchant\protector\apply-magic-gloves.cpp" />
+    <ClCompile Include="..\..\src\object-enchant\protector\apply-magic-helm.cpp" />
+    <ClCompile Include="..\..\src\object-enchant\others\apply-magic-ring.cpp" />
+    <ClCompile Include="..\..\src\object-enchant\protector\apply-magic-shield.cpp" />
     <ClCompile Include="..\..\src\object\object-index-list.cpp" />
     <ClCompile Include="..\..\src\player-info\alignment.cpp" />
     <ClCompile Include="..\..\src\player-info\equipment-info.cpp" />
@@ -665,9 +665,9 @@
     <ClCompile Include="..\..\src\mind\mind-samurai.cpp" />
     <ClCompile Include="..\..\src\mind\mind-warrior.cpp" />
     <ClCompile Include="..\..\src\racial\racial-vampire.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\apply-magic-armor.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\apply-magic-others.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\apply-magic-weapon.cpp" />
+    <ClCompile Include="..\..\src\object-enchant\protector\apply-magic-armor.cpp" />
+    <ClCompile Include="..\..\src\object-enchant\others\apply-magic-others.cpp" />
+    <ClCompile Include="..\..\src\object-enchant\weapon\apply-magic-weapon.cpp" />
     <ClCompile Include="..\..\src\object-enchant\apply-magic.cpp" />
     <ClCompile Include="..\..\src\player\eldritch-horror.cpp" />
     <ClCompile Include="..\..\src\specific-object\bow.cpp" />
@@ -938,11 +938,11 @@
     <ClInclude Include="..\..\src\monster-race\monster-aura-types.h" />
     <ClInclude Include="..\..\src\monster-race\race-behavior-flags.h" />
     <ClInclude Include="..\..\src\mspell\mspell-result.h" />
-    <ClInclude Include="..\..\src\object-enchant\abstract-weapon-enchanter.h" />
-    <ClInclude Include="..\..\src\object-enchant\object-smith.h" />
-    <ClInclude Include="..\..\src\object-enchant\smith-info.h" />
-    <ClInclude Include="..\..\src\object-enchant\smith-tables.h" />
-    <ClInclude Include="..\..\src\object-enchant\smith-types.h" />
+    <ClInclude Include="..\..\src\object-enchant\weapon\abstract-weapon-enchanter.h" />
+    <ClInclude Include="..\..\src\smith\object-smith.h" />
+    <ClInclude Include="..\..\src\smith\smith-info.h" />
+    <ClInclude Include="..\..\src\smith\smith-tables.h" />
+    <ClInclude Include="..\..\src\smith\smith-types.h" />
     <ClInclude Include="..\..\src\object-enchant\tr-flags.h" />
     <ClInclude Include="..\..\src\object-use\item-use-checker.h" />
     <ClInclude Include="..\..\src\object-use\throw-execution.h" />
@@ -1000,15 +1000,15 @@
     <ClInclude Include="..\..\src\main-win\main-win-term.h" />
     <ClInclude Include="..\..\src\main-win\wav-reader.h" />
     <ClInclude Include="..\..\src\monster\monster-damage.h" />
-    <ClInclude Include="..\..\src\object-enchant\apply-magic-amulet.h" />
-    <ClInclude Include="..\..\src\object-enchant\abstract-protector-enchanter.h" />
-    <ClInclude Include="..\..\src\object-enchant\apply-magic-boots.h" />
-    <ClInclude Include="..\..\src\object-enchant\apply-magic-cloak.h" />
-    <ClInclude Include="..\..\src\object-enchant\apply-magic-crown.h" />
-    <ClInclude Include="..\..\src\object-enchant\apply-magic-gloves.h" />
-    <ClInclude Include="..\..\src\object-enchant\apply-magic-helm.h" />
-    <ClInclude Include="..\..\src\object-enchant\apply-magic-ring.h" />
-    <ClInclude Include="..\..\src\object-enchant\apply-magic-shield.h" />
+    <ClInclude Include="..\..\src\object-enchant\others\apply-magic-amulet.h" />
+    <ClInclude Include="..\..\src\object-enchant\protector\abstract-protector-enchanter.h" />
+    <ClInclude Include="..\..\src\object-enchant\protector\apply-magic-boots.h" />
+    <ClInclude Include="..\..\src\object-enchant\protector\apply-magic-cloak.h" />
+    <ClInclude Include="..\..\src\object-enchant\protector\apply-magic-crown.h" />
+    <ClInclude Include="..\..\src\object-enchant\protector\apply-magic-gloves.h" />
+    <ClInclude Include="..\..\src\object-enchant\protector\apply-magic-helm.h" />
+    <ClInclude Include="..\..\src\object-enchant\others\apply-magic-ring.h" />
+    <ClInclude Include="..\..\src\object-enchant\protector\apply-magic-shield.h" />
     <ClInclude Include="..\..\src\object-use\use-execution.h" />
     <ClInclude Include="..\..\src\object-use\zaprod-execution.h" />
     <ClInclude Include="..\..\src\object-use\zapwand-execution.h" />
@@ -1439,9 +1439,9 @@
     <ClInclude Include="..\..\src\mind\mind-warrior.h" />
     <ClInclude Include="..\..\src\racial\racial-vampire.h" />
     <ClInclude Include="..\..\src\object-enchant\enchanter-base.h" />
-    <ClInclude Include="..\..\src\object-enchant\apply-magic-armor.h" />
-    <ClInclude Include="..\..\src\object-enchant\apply-magic-others.h" />
-    <ClInclude Include="..\..\src\object-enchant\apply-magic-weapon.h" />
+    <ClInclude Include="..\..\src\object-enchant\protector\apply-magic-armor.h" />
+    <ClInclude Include="..\..\src\object-enchant\others\apply-magic-others.h" />
+    <ClInclude Include="..\..\src\object-enchant\weapon\apply-magic-weapon.h" />
     <ClInclude Include="..\..\src\object-enchant\apply-magic.h" />
     <ClInclude Include="..\..\src\player\eldritch-horror.h" />
     <ClInclude Include="..\..\src\realm\realm-types.h" />
