+++ src/artifact/artifact-info.cpp
@@ -7,9 +7,9 @@
 #include "artifact/random-art-effects.h"
 #include "object-enchant/activation-info-table.h"
 #include "object-enchant/object-ego.h"
-#include "object-enchant/object-smith.h"
 #include "object-enchant/tr-types.h"
 #include "object/object-kind.h"
+#include "smith/object-smith.h"
 #include "system/artifact-type-definition.h"
 #include "system/object-type-definition.h"
 #include "util/bit-flags-calculator.h"
