+++ src/io-dump/special-class-dump.cpp
@@ -10,12 +10,12 @@
 #include "mind/mind-blue-mage.h"
 #include "monster-race/race-ability-flags.h"
 #include "mspell/monster-power-table.h"
-#include "object-enchant/object-smith.h"
 #include "object/object-kind-hook.h"
 #include "object/object-kind.h"
 #include "player-base/player-class.h"
 #include "player-info/bluemage-data-type.h"
 #include "player-info/magic-eater-data-type.h"
+#include "smith/object-smith.h"
 #include "system/player-type-definition.h"
 #include "util/enum-converter.h"
 #include "util/flag-group.h"
