+++ src/object-enchant/others/apply-magic-amulet.h
