+++ src/object-enchant/others/apply-magic-others.h
