+++ src/object-enchant/others/apply-magic-ring.h
