+++ src/object-enchant/protector/abstract-protector-enchanter.h
