+++ src/object-enchant/protector/apply-magic-boots.h
@@ -1,20 +1,20 @@
 ﻿#pragma once
 
-#include "object-enchant/abstract-protector-enchanter.h"
+#include "object-enchant/protector/abstract-protector-enchanter.h"
 #include "object-enchant/enchanter-base.h"
 #include "system/angband.h"
 
 struct object_type;
 class PlayerType;
-class ShieldEnchanter : AbstractProtectorEnchanter {
+class BootsEnchanter : AbstractProtectorEnchanter {
 public:
-    ShieldEnchanter(PlayerType *player_ptr, object_type *o_ptr, DEPTH level, int power);
-    virtual ~ShieldEnchanter() = default;
+    BootsEnchanter(PlayerType *player_ptr, object_type *o_ptr, DEPTH level, int power);
+    virtual ~BootsEnchanter() = default;
     void apply_magic() override;
 
 protected:
     void enchant() override{};
-    void give_ego_index() override;
+    void give_ego_index() override{};
     void give_high_ego_index() override{};
     void give_cursed() override{};
