+++ src/object-enchant/protector/apply-magic-cloak.cpp
@@ -1,14 +1,12 @@
 ﻿/*
- * @brief 籠手に耐性等の追加効果を付与する処理
+ * @brief クロークに耐性等の追加効果を付与する処理
  * @date 2021/08/01
  * @author Hourier
- * @details ドラゴングローヴは必ず付与する. それ以外は確率的に付与する.
  */
 
-#include "object-enchant/apply-magic-gloves.h"
+#include "object-enchant/protector/apply-magic-cloak.h"
 #include "artifact/random-art-generator.h"
 #include "inventory/inventory-slot-types.h"
-#include "object-enchant/object-boost.h"
 #include "object-enchant/object-ego.h"
 #include "sv-definition/sv-protector-types.h"
 #include "system/object-type-definition.h"
@@ -20,7 +18,7 @@
  * @param level 生成基準階
  * @param power 生成ランク
  */
-GlovesEnchanter::GlovesEnchanter(PlayerType *player_ptr, object_type *o_ptr, DEPTH level, int power)
+CloakEnchanter::CloakEnchanter(PlayerType *player_ptr, object_type *o_ptr, DEPTH level, int power)
     : AbstractProtectorEnchanter{ o_ptr, level, power }
     , player_ptr(player_ptr)
 {
@@ -29,13 +27,10 @@ GlovesEnchanter::GlovesEnchanter(PlayerType *player_ptr, object_type *o_ptr, DEP
 /*
  * @details power > 2はデバッグ専用.
  */
-void GlovesEnchanter::apply_magic()
+void CloakEnchanter::apply_magic()
 {
-    if (this->o_ptr->sval == SV_SET_OF_DRAGON_GLOVES) {
-        dragon_resist(this->o_ptr);
-        if (!one_in_(3)) {
-            return;
-        }
+    if (this->o_ptr->sval == SV_ELVEN_CLOAK) {
+        this->o_ptr->pval = randint1(4);
     }
 
     if (this->power > 1) {
@@ -44,11 +39,11 @@ void GlovesEnchanter::apply_magic()
             return;
         }
 
-        this->o_ptr->name2 = get_random_ego(INVEN_ARMS, true);
+        this->o_ptr->name2 = get_random_ego(INVEN_OUTER, true);
         return;
     }
-    
+
     if (this->power < -1) {
-        this->o_ptr->name2 = get_random_ego(INVEN_ARMS, false);
+        this->o_ptr->name2 = get_random_ego(INVEN_OUTER, false);
     }
 }
