+++ src/object-enchant/protector/apply-magic-helm.cpp
@@ -1,14 +1,16 @@
 ﻿/*
- * @brief 冠に耐性等の追加効果を付与する処理
+ * @brief 兜に耐性等の追加効果を付与する処理
  * @date 2021/08/01
  * @author Hourier
+ * @details ドラゴンヘルムは必ず付与する. それ以外は確率的に付与する.
  */
 
-#include "object-enchant/apply-magic-crown.h"
+#include "object-enchant/protector/apply-magic-helm.h"
 #include "artifact/random-art-generator.h"
 #include "inventory/inventory-slot-types.h"
 #include "object-enchant/object-boost.h"
 #include "object-enchant/object-ego.h"
+#include "sv-definition/sv-protector-types.h"
 #include "system/object-type-definition.h"
 
 /*
@@ -18,14 +20,21 @@
  * @param level 生成基準階
  * @param power 生成ランク
  */
-CrownEnchanter::CrownEnchanter(PlayerType *player_ptr, object_type *o_ptr, DEPTH level, int power)
+HelmEnchanter::HelmEnchanter(PlayerType *player_ptr, object_type *o_ptr, DEPTH level, int power)
     : AbstractProtectorEnchanter{ o_ptr, level, power }
     , player_ptr(player_ptr)
 {
 }
 
-void CrownEnchanter::apply_magic()
+void HelmEnchanter::apply_magic()
 {
+    if (this->o_ptr->sval == SV_DRAGON_HELM) {
+        dragon_resist(this->o_ptr);
+        if (!one_in_(3)) {
+            return;
+        }
+    }
+
     if (this->power > 1) {
         this->give_ego_index();
         return;
@@ -39,7 +48,7 @@ void CrownEnchanter::apply_magic()
 /*
  * @details power > 2はデバッグ専用.
  */
-void CrownEnchanter::give_ego_index()
+void HelmEnchanter::give_ego_index()
 {
     if (one_in_(20) || (this->power > 2)) {
         become_random_artifact(this->player_ptr, this->o_ptr, false);
@@ -49,15 +58,14 @@ void CrownEnchanter::give_ego_index()
     while (true) {
         this->o_ptr->name2 = get_random_ego(INVEN_HEAD, true);
         switch (this->o_ptr->name2) {
-        case EGO_TELEPATHY:
-        case EGO_MAGI:
-        case EGO_MIGHT:
-        case EGO_REGENERATION:
-        case EGO_LORDLINESS:
-        case EGO_BASILISK:
+        case EGO_BRILLIANCE:
+        case EGO_DARK:
+        case EGO_INFRAVISION:
+        case EGO_H_PROTECTION:
+        case EGO_LITE:
             return;
         case EGO_SEEING:
-            if (one_in_(3)) {
+            if (one_in_(7)) {
                 add_low_telepathy(this->o_ptr);
             }
 
@@ -68,12 +76,12 @@ void CrownEnchanter::give_ego_index()
     }
 }
 
-void CrownEnchanter::give_cursed()
+void HelmEnchanter::give_cursed()
 {
     while (true) {
         this->o_ptr->name2 = get_random_ego(INVEN_HEAD, false);
         switch (this->o_ptr->name2) {
-        case EGO_H_DEMON:
+        case EGO_ANCIENT_CURSE:
             return;
         default:
             continue;
