+++ src/object-enchant/protector/apply-magic-shield.h
@@ -1,20 +1,20 @@
 ﻿#pragma once
 
-#include "object-enchant/abstract-protector-enchanter.h"
+#include "object-enchant/protector/abstract-protector-enchanter.h"
 #include "object-enchant/enchanter-base.h"
 #include "system/angband.h"
 
 struct object_type;
 class PlayerType;
-class GlovesEnchanter : AbstractProtectorEnchanter {
+class ShieldEnchanter : AbstractProtectorEnchanter {
 public:
-    GlovesEnchanter(PlayerType *player_ptr, object_type *o_ptr, DEPTH level, int power);
-    virtual ~GlovesEnchanter() = default;
+    ShieldEnchanter(PlayerType *player_ptr, object_type *o_ptr, DEPTH level, int power);
+    virtual ~ShieldEnchanter() = default;
     void apply_magic() override;
 
 protected:
     void enchant() override{};
-    void give_ego_index() override{};
+    void give_ego_index() override;
     void give_high_ego_index() override{};
     void give_cursed() override{};
