+++ src/object/object-flags.cpp
@@ -1,10 +1,10 @@
 ﻿#include "object/object-flags.h"
 #include "mind/mind-weaponsmith.h"
 #include "object-enchant/object-ego.h"
-#include "object-enchant/object-smith.h"
 #include "object-enchant/tr-types.h"
 #include "object/object-kind.h"
 #include "perception/object-perception.h"
+#include "smith/object-smith.h"
 #include "sv-definition/sv-lite-types.h"
 #include "system/artifact-type-definition.h"
 #include "system/object-type-definition.h"
