+++ src/smith/object-smith.h
