+++ src/smith/smith-info.h
