+++ src/smith/smith-tables.h
