+++ src/wizard/wizard-spells.cpp
@@ -16,13 +16,13 @@
 #include "monster-floor/monster-generator.h"
 #include "monster-floor/monster-summon.h"
 #include "monster-floor/place-monster-types.h"
-#include "monster-race/race-ability-flags.h"
 #include "monster-race/monster-race.h"
+#include "monster-race/race-ability-flags.h"
 #include "mutation/mutation-processor.h"
-#include "object-enchant/object-smith.h"
 #include "player-base/player-class.h"
 #include "player-info/bluemage-data-type.h"
 #include "player-info/smith-data-type.h"
+#include "smith/object-smith.h"
 #include "spell-kind/spells-launcher.h"
 #include "spell-kind/spells-random.h"
 #include "spell-kind/spells-teleport.h"
@@ -30,8 +30,8 @@
 #include "spell/spells-status.h"
 #include "spell/summon-types.h"
 #include "system/floor-type-definition.h"
-#include "system/player-type-definition.h"
 #include "system/monster-race-definition.h"
+#include "system/player-type-definition.h"
 #include "target/grid-selector.h"
 #include "target/target-checker.h"
 #include "target/target-getter.h"
@@ -193,7 +193,7 @@ void wiz_summon_specific_enemy(PlayerType *player_ptr, MONRACE_IDX r_idx)
 {
     if (r_idx <= 0) {
         int val = 1;
-        if(!get_value("MonsterID", 1, r_info.size() - 1, &val)) {
+        if (!get_value("MonsterID", 1, r_info.size() - 1, &val)) {
             return;
         }
         r_idx = static_cast<MONRACE_IDX>(val);
